src/service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
datasets==4.8.4
dill==0.4.1
fastapi==0.136.0
filelock==3.29.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.11.0
hyperframe==6.1.0
idna==3.12
importlib_metadata==8.7.1
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.4.2
numpy==2.2.6
onnxruntime==1.23.2
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-instrumentation==0.62b0
opentelemetry-instrumentation-asgi==0.62b0
opentelemetry-instrumentation-fastapi==0.62b0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
opentelemetry-util-http==0.62b0
packaging==26.1
pandas==2.3.3
portalocker==3.2.0
prometheus_client==0.25.0
propcache==0.4.1
protobuf==6.33.6
pyarrow==24.0.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
qdrant-client==1.17.1
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.45.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.1
celery==5.3.1
redis==4.6.0
dvc==3.63.0